refac! move identity features list to utils/collections
